Restricted Pages

+++ base/views.py
@@ -5,6 +5,8 @@ from .forms import RoomForm
 from django.contrib.auth.models import User
 from django.contrib import messages
 from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.decorators import login_required
+from django.http import HttpResponse
 
 
 # Python key value pair
@@ -16,23 +18,22 @@ from django.contrib.auth import authenticate, login, logout
 
 def loginPage(request):
 
+    if request.user.is_authenticated:
+        return redirect('home')
+
     if request.method == 'POST':
         username = request.POST.get('username')
         password = request.POST.get('password')
-
         try:
             user = User.objects.get(username=username)
         except:
             messages.error(request, 'Username does not exist')
-
         user = authenticate(request, username=username, password=password)
-        
         if user is not None:
             login(request, user) #This logs the user in and creates a session
             return redirect('home')
         else:
             messages.error(request, 'Username OR Password does not exist')
-
     context = {}
     return render(request, 'base/login_register.html', context)
 
@@ -47,13 +48,11 @@ def logoutUser(request):
 # The context var is passed to the home.html site and displayed on the page by using a for loop
 def home(request):
     q = request.GET.get('q') if request.GET.get('q') != None else '' #q is equal to whatever we pass into the url
-    
     rooms = Room.objects.filter(  #Check if a object contain the value in q and set that to 'room', the Q allows us to check if q is in multiple fields
         Q(topic__name__icontains=q) |
         Q(name__icontains=q) |
         Q(description__icontains=q)
         )
-    
     topics = Topic.objects.all()
     room_count = rooms.count() #Gives us the number of rooms
     context = {'rooms': rooms, 'topics': topics, 'room_count': room_count}
@@ -67,6 +66,7 @@ def room(request, pk):
 
 
 #Creates a new room and saves it if the input is valid
+@login_required(login_url='login') #Requires user to be logged in to access this function, if not the user is redirected to the login page
 def createRoom(request):
     form = RoomForm()
     if request.method == 'POST':
@@ -80,20 +80,25 @@ def createRoom(request):
 
 #Method for updating the content of a room
 #We enter the primary key to know which item we are updating
+@login_required(login_url='login') #Requires user to be logged in to access this function, if not the user is redirected to the login page
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk) #What is being updated
     form = RoomForm(instance=room) #Creates a RoomForm instance with the prefilled values from the db
+    if request.user != room.host: #Check if user is the creator of the room, if not the httpresponse will be triggered
+        return HttpResponse('You are not allowed here')
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room) #Updates the values of the form under this id
         if form.is_valid():
             form.save() # Saves the values to the db under the id
-            return redirect('home') #Returns the user to the home page      
+            return redirect('home') #Returns the user to the home page          
     context = {'form': form}
     return render(request, 'base/room_form.html', context) #Context is what is being sent to the html page
 
 
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+    if request.user != room.host:
+        return HttpResponse('You are not allowed here')
     if request.method == 'POST':
         room.delete()
         return redirect('home')
